% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Address_Parser.R
\name{Address_Parser}
\alias{Address_Parser}
\title{Parses an Address}
\usage{
Address_Parser(address)
}
\arguments{
\item{address}{address the address as a character to be parsed.}
}
\value{
splt_addr a data frame with columns street, city,
state abbreviation, county, and state
}
\description{
takes in an address string separated by commas and returns the
address parsed in list form.
}
\author{
Alex Cory
}
