% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_Capita_Grocery_Spend.R
\name{Adj_Capita_Grocery_Spend}
\alias{Adj_Capita_Grocery_Spend}
\title{Calculates Average Per Capita Spending Adjusted by Estimated Price Increase}
\usage{
Adj_Capita_Grocery_Spend(
  est_per_price_increase = 7,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{est_per_price_increase}{Estimated increase in price of grocery in
cumulative percent default taken as 7 for 2023}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
Outputs the average per capita grocery spend after adjusting for the expected price increase.
}
\description{
This function returns the average per capita spending after adjusting for
percentage price increase from the base year(2022).
}
\details{
This function is taken from the Estimating_Revenue.xlsx and calculated in Step two.
This function calls Avg_Capita_Grocery_Spending().
}
\examples{
Adj_Capita_Grocery_Spend()

}
\author{
Srika Raja
}
