% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annual_Rent.R
\name{Annual_Rent}
\alias{Annual_Rent}
\title{Calculation of Annual Rent}
\usage{
Annual_Rent(Monthly_Rent)
}
\arguments{
\item{Monthly_Rent}{Monthly rent for the lease}
}
\value{
Outputs Annual Rent (in dollars)
}
\description{
Calculation of Annual Rent
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and calculated in Step Four
}
\examples{
# Annual_Rent(Monthly_Building_Rent, Other_Rent_1, Other_Rent_2)

}
