% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Pct.R
\name{Auto_Pct}
\alias{Auto_Pct}
\title{Determining Percentage of County Covered in Market Area}
\usage{
Auto_Pct(
  northeast_dist,
  northwest_dist,
  southeast_dist,
  southwest_dist,
  df_city_pop,
  geo_county
)
}
\arguments{
\item{northeast_dist}{Distance to nearest store to the Northeast}

\item{northwest_dist}{Distance to nearest store to the Northwest}

\item{southeast_dist}{Distance to nearest store to the Southeast}

\item{southwest_dist}{Distance to nearest store to the Southwest}

\item{df_city_pop}{Dataframe with City Population}

\item{geo_county}{County Name}
}
\value{
Sum of Market Area Divided by Size of County
}
\description{
Calculates total quadrant size.
Determines size and percentage of County covered by Market Area.
}
\author{
Alex Cory
}
