% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_Rural_Pop.R
\name{Auto_Rural_Pop}
\alias{Auto_Rural_Pop}
\title{Calculates the Rural Population for a Given Market Size}
\usage{
Auto_Rural_Pop(df_buffer_loc)
}
\arguments{
\item{df_buffer_loc}{a data frame containing the city, county and state
locations of a buffered point.Inherited from the returned df_census_call
data frame returned in the call DSPGrocery::Create_Circle_Buffer.}
}
\value{
the function returns an estimated number for the rural population in
the calculated market size.
}
\description{
This is a helper function for DSPGrocery::FILLTHISLATER to calculate the
rural population for the buffered geographic area using
DSPGrocery::Create_Circle_Buffer.The calculation for this is the difference between all county
populations and city populations in those counties multiplied by the percentage
of area that the buffer covers in intersecting counties.
}
\details{
This function calls DSPGrocery::Calc_Cities_Pop and
DSPGrocery::Calc_Cities_Pop
}
\author{
Harun Celik
}
