% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Avg_Capita_Grocery_Spending.R
\name{Avg_Capita_Grocery_Spending}
\alias{Avg_Capita_Grocery_Spending}
\title{Calculates Average Grocery Spending Per Capita.}
\usage{
Avg_Capita_Grocery_Spending(
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
Outputs average grocery spending per capita.
}
\description{
This function returns the average per capita spending by taking the ratio of
grocery sales to population in the base year(2022).
}
\details{
This function is taken from the Estimating_Revenue.xlsx and calculated in Step One.
}
\examples{
Avg_Capita_Grocery_Spending(grocery_sales = 811541000000, population = 334233854)

}
\author{
Srika Raja
}
