% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Cities_Pop.R
\name{Calc_Cities_Pop}
\alias{Calc_Cities_Pop}
\title{Calculates the Population of Cities by County Geography}
\usage{
Calc_Cities_Pop(states_unique, all_counties_pops)
}
\arguments{
\item{states_unique}{this is a vector of unique state FIPS codes inherited from
the df_buffer_loc parameter in DSPGrocery::Auto_Rural_Pop.}

\item{all_counties_pops}{this is the returned data frame inherited from the
call DSPGrocery::Calc_Counties_Pop.}
}
\value{
the function returns a data frame called \code{all_cities_intersect} containing
county, city, city population, and sf geographies. The data frame is returned
with an attribute called "all_population" which is the sum of city populations.
}
\description{
This is a helper function for the DSPGrocery::Auto_Rural_Pop function which
calculates the total population of cities within a county that
intersect county geography. The call first pulls all cities in a provided state
and uses an sf::st_join to join based on county instersections.
}
\details{
The call uses data pulled from the United States Census Bureau with a call to
the tidycensus API, tidycensus::get_decennial.
}
\seealso{
\code{\link[=Calc_Counties_Pop]{Calc_Counties_Pop()}}
}
\author{
Harun Celik
}
