% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Counties_Pop.R
\name{Calc_Counties_Pop}
\alias{Calc_Counties_Pop}
\title{Calculates the Population of Counties Given a Vector of Counties and States}
\usage{
Calc_Counties_Pop(states_unique, df_buffer_loc)
}
\arguments{
\item{states_unique}{this is a vector of unique state FIPS codes inherited from
the df_buffer_loc parameter in DSPGrocery::Auto_Rural_Pop.}

\item{df_buffer_loc}{this is a data frame inherited from DSPGrocery::Auto_Rural_Pop.}
}
\value{
the function returns a data frame called \code{all_counties_pops} containing
county, county population, and sf geographies. The data frame is returned
with an attribute called "all_population" which is the sum of county populations.
}
\description{
This is a helper function for the DSPGrocery::Auto_Rural_Pop function which
calculates the total population of counties within states given a data frame
containing a vector of county names and state names.
}
\details{
The call uses data pulled from the United States Census Bureau with a call to
the tidycensus API, tidycensus::get_decennial.
}
\seealso{
\code{\link[=Calc_Cities_Pop]{Calc_Cities_Pop()}}
}
\author{
Harun Celik
}
