% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Market_Size.R
\name{Calc_Market_Size}
\alias{Calc_Market_Size}
\title{Calculates Market Size}
\usage{
Calc_Market_Size(address)
}
\arguments{
\item{address}{an address separated by commas containing street, city, and abbreviated state.
Example: "Main Street, Lamoni, IA"}
}
\value{
returns a list of atomic vectors containing city_population, metro_population,
and rural_population.
}
\description{
this function automatically calculates the market size for a given address value.
}
\details{
this function calls DSPGGrocery::Pop_Binder, DSPGGrocery::Cities_Pop, DSPGGrocery::Metro_Pop
and DSPGGrocery::Rural_Pop.
}
\author{
Alex Cory
}
