% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cities_Pop.R
\name{Cities_Pop}
\alias{Cities_Pop}
\title{Finds the Population of all the Cities in the Market Area}
\usage{
Cities_Pop(df_city_state, city_county_state)
}
\arguments{
\item{df_city_state}{df from Pop_Binder with cities and populations}

\item{city_county_state}{df with city county and state columns}
}
\value{
Population of all towns within region
}
\description{
Determines and cleans all the cities in the market area.
Finds the sum of all of their populations.
}
\details{
Needs Tidycensus key
}
\author{
Alex Cory
}
