% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/City_Populations.R
\name{City_Populations}
\alias{City_Populations}
\title{Calculation for population of the cities in region.}
\usage{
City_Populations(list_in)
}
\arguments{
\item{list}{user defined list of the population of towns}
}
\value{
outputs sum of town populations
}
\description{
Calculation for population of the cities in region.
}
\examples{
City_Populations(list(77))
}
