% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/City_Populations.R
\name{CityPopulations}
\alias{CityPopulations}
\title{Calculation for a quarter circle.}
\usage{
CityPopulations(list)
}
\arguments{
\item{list}{user defined list of the population of towns}
}
\value{
outputs sum of town populations
}
\description{
Calculation for a quarter circle.
}
\examples{
CityPopulations(list)
}
