% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cost_of_Goods_Sold.R
\name{Cost_of_Goods_Sold}
\alias{Cost_of_Goods_Sold}
\title{Calculation of the Cost of Goods Sold}
\usage{
Cost_of_Goods_Sold(Total_Estimated_Revenue, Gross_Margin_Percentage = 0.2446)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of a total estimated revenue calculation}

\item{Gross_Margin_Percentage}{Percentage of revenue as gross margin (default from Bizminer)}
}
\value{
Outputs cost of goods sold (in dollars)
}
\description{
Calculation of the Cost of Goods Sold
}
\details{
Calls Gross_Margin() function to provide the gross margin (in dollars)
which is subtracted from revenue to
calculate cost of goods.
This function was taken from Estimating_Expense.xlxs (Step 6)
}
\examples{
# Cost_of_Goods_Sold(Total_Estimated_Revenue = 2000000, Gross_Margin_Percentage = .2446)

}
