% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/County_Identifier.R
\name{County_Identifier}
\alias{County_Identifier}
\title{Find County from Long/Lat points and state}
\usage{
County_Identifier(x)
}
\arguments{
\item{x}{Single row data frame,columns 1, 2 = lon, lat, column 3 = state name}
}
\value{
x with extra column for county name
}
\description{
Uses Latitude, Longitude, and State to determines the county.
}
\author{
Jay Maxwell
}
