% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_Circle_Buffer.R
\name{Create_Circle_Buffer}
\alias{Create_Circle_Buffer}
\title{Creates a Circle Buffer to the Furthest of Four Nearest Stores in a Quadrant}
\usage{
Create_Circle_Buffer(address, api_key, keyword)
}
\arguments{
\item{address}{a comma seperated address as a character.}

\item{api_key}{a google api key to be used as the key argument in googleway::set_key.}

\item{keyword}{a keyword argument used by the googleway::google_places api call.}
}
\value{
The output returns the spatial features required to create the leaflet map.
the function returns a list called Store_Info which holds the objects
df_grocery_all, df_grocery_only, df_grocery_circle, buffer_point_origin,
store_points, furthest_point,buffer_circle, all_counties, grocery_counties_inter,
grocery_cities_inter
}
\description{
This function produces the data frames required for the spatial mapping and
census api calls in the DSPGrocery tool. See the returns section to check the
list of all objects.
}
\details{
The function makes use of the googleway and the tigris API. Only the googleway
API requires a key.
}
\author{
Harun Celik
}
