% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depreciation_1.R
\name{Depreciation_1}
\alias{Depreciation_1}
\title{Calculation of Depreciation (for Building Ownership)}
\usage{
Depreciation_1(
  Building_Remodeling,
  Parking_Lot_Improvements,
  Shelving_Check_Out_Counters,
  Computer_Equipment_POS,
  Vehicles,
  Display_Cases,
  Refrigeration,
  Freezers,
  Meat_Cutting_Equipment,
  Miscellaneous_Assets_1,
  Miscellaneous_Assets_1_Use_Life,
  Miscellaneous_Assets_2,
  Miscellaneous_Assets_2_Use_Life,
  Miscellaneous_Assets_3,
  Miscellaneous_Assets_3_Use_Life
)
}
\value{
Outputs total estimated loss from depreciation for property owners
}
\description{
Calculation of Depreciation (for Building Ownership)
}
\details{
This function is taken from the "Estimating Expense" excel sheet (step two)
}
