% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distance_Euclidean.R
\name{Distance_Euclidean}
\alias{Distance_Euclidean}
\title{Calculates Euclidean Distance from an Origin point to End Points}
\usage{
Distance_Euclidean(df, origin, end)
}
\arguments{
\item{df}{this is the data frame containing the latitude and longitude points
to which the distance column will be binded to.}

\item{origin}{this is a vector containing a single latitude and longitude point
to be used as the origin point.}

\item{end}{this is a matrix of latitude and longitude points to be used as the
end points.}
}
\value{
The function returns df_new, a data frame that has the distance calculations
appended in the distance_vector column.
}
\description{
The function calculates the euclidean distance from a provided origin
point to provided end points. The origin and param parameters should be provided
in a vector format.
}
\author{
Harun Celik
}
