% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Employee_Wages.R
\name{Employee_Wages}
\alias{Employee_Wages}
\title{Calculation of Employee Wages}
\usage{
Employee_Wages(Total_Estimated_Revenue)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of total estimated revenue calculation sheet}
}
\value{
Outputs dollars spent on employee wages (in dollars based off of percentages taken from Bizminer)
}
\description{
Calculation of Employee Wages
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and calculated in Step Six)
}
\examples{
# Employee_Wages(Total_Estimated_Revenue = 2000000)

}
