% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Census_Vars_Decennial.R
\name{Get_Census_Vars_Decennial}
\alias{Get_Census_Vars_Decennial}
\title{Retrieves Relevant Census Variables from the US Decennial Census (2020)}
\usage{
Get_Census_Vars_Decennial(df_locations)
}
\arguments{
\item{df_locations}{Data frame of cities, counties and states within the range of a circle buffer
determined by the location of nearest grocery stores. This data frame is inherited
from the return of DSPGrocery::Create_Circle_Buffer as df_census_call.}
}
\value{
The output returns a data frame of the 2020 decennial census variables
for race/ethnicity and total population for each county.
}
\description{
This function retrieves the variables related to race/ethnicity and total population
for each county from the U.S Decennial Census (2020). These variables are then used to
render a data table and a plot in the tool.
}
\details{
This function receives the data frame of cities, counties and states within the
circle buffer established by the market size spatial calculations and outputs census data
for each county.
}
\author{
Aaron Null
}
