% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Cities_In_County.R
\name{Get_Cities_in_County}
\alias{Get_Cities_in_County}
\title{Find all Cities in a County Given State and County}
\usage{
Get_Cities_in_County(x)
}
\arguments{
\item{x}{a single-row data frame, a column named "state" a column named "county"}
}
\value{
x with all cities in a column as a list
}
\description{
Find all Cities in a County Given State and County
}
\author{
Jay Maxwell
}
