% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Haversine_Calculator.R
\name{Haversine_Calculator}
\alias{Haversine_Calculator}
\title{Find Distance using Haversine Formula}
\usage{
Haversine_Calculator(df_grocery_only, df_geocode)
}
\arguments{
\item{df_grocery_only}{dataframe with API call of grocery stores}

\item{df_geocode}{dataframe with gps coordinates of store}
}
\value{
Distance_Comp_List with nearest store in each direction and
the latlong information on those stores
}
\description{
Finds the distance between a point and a dataframe of points using
the Haversine formula, and finds nearest stores in each direction.
Uses Haversine Formula to account for the curvature of the Earth.
}
\author{
Alex Cory
}
