% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interest_Expense.R
\name{Interest_Expense}
\alias{Interest_Expense}
\title{Calculation of first-year interest on a loan}
\usage{
Interest_Expense(Loan_Amount, Interest_Rate)
}
\arguments{
\item{Loan_Amount}{Total dollar amount for loan}

\item{Interest_Rate}{Interest rate (decimal form) on the loan}
}
\value{
Outputs first-year loan interest (in dollars)
}
\description{
Calculation of first-year interest on a loan
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and calculated in Step Five.
}
\examples{
# Interest_Expense(Loan_Amount = 1000000, Interest_Rate = .06)

}
