% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metro_Pop.R
\name{Metro_Pop}
\alias{Metro_Pop}
\title{Finds the Population of the City the Store is in.}
\usage{
Metro_Pop(address, df_city_pop)
}
\arguments{
\item{address}{the address we are interested in.}

\item{df_city_pop}{dataframe with population of cities in county}
}
\value{
population of city the store is in.
}
\description{
Returns the population of the city you are currently in
}
\author{
Alex Cory
}
