% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Officer_Compensation.R
\name{Officer_Compensation}
\alias{Officer_Compensation}
\title{Calculation of Officer Compensation}
\usage{
Officer_Compensation(Total_Estimated_Revenue, Percentage = 0.0158)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of total estimated revenue calculation sheet}

\item{Percentage}{Percentage of total estimated revenue spent on officer compensation (taken from Bizminer)}
}
\value{
Outputs dollars spent (Total Estimated Revenue * Percentage) (in dollars)
}
\description{
Calculation of Officer Compensation
}
\examples{
# Officer_Compensation(Total_Estimated_Revenue = 2000000, Percentage = .0158)

}
