% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operating_Income_Loss.R
\name{Operating_Income_Loss}
\alias{Operating_Income_Loss}
\title{Calculation of the operating income (loss)}
\usage{
Operating_Income_Loss(Total_Estimated_Revenue)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of total estimated revenue calculation sheet}
}
\value{
Outputs difference between total estimated revenue and aggregated expenses (in dollars)
}
\description{
Calculation of the operating income (loss)
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and calculated in Step Six)
Functions called: Cost_of_Goods_Sold, Officer_Compensation, Employee_Wages, Other_Operating_Expense
}
\examples{
# Employee_Wages(Total_Estimated_Revenue = 2000000)

}
