% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Income.R
\name{Other_Income}
\alias{Other_Income}
\title{Calculates the Estimated Income from Miscellaneous Sources for Grocery Stores}
\usage{
Other_Income(Total_Estimated_Revenue, Percentage = 0.0115)
}
\arguments{
\item{Total_Estimated_Revenue}{Total estimated revenue based on market size/location.}

\item{Percentage}{Percentage of total estimated revenue that constitutes the added
income from miscellaneous ("other") income sources selected via user input
(default from Bizminer).}
}
\value{
The output returns the estimated dollar amount earned as income from these other
sources of income for a hypothetical grocery store in a given location.
}
\description{
This function calculates the estimated income from miscellaneous income sources
for a hypothetical store in a user-selected location based off of a user-selected
percentage parameter. If not specified by the user, this percentage defaults
to the industry average provided by Bizminer. Some examples of miscellaneous sources
of income might include special grocery delivery services, curbside pickup or discount
club fees.
}
\details{
This function employs a default percentage value from the financial analysis
and market research firm Bizminer and is based upon calculations originally
formulated by FFED ISU Extension and Outreach.
}
\author{
Aaron Null
}
