% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primary_Secondary_Shoppers.R
\name{Primary_Secondary_Shoppers}
\alias{Primary_Secondary_Shoppers}
\title{Calculation for a quarter circle.}
\usage{
Primary_Secondary_Shoppers(pop, pct_primary, pct_secondary)
}
\arguments{
\item{pop}{population inside quadrants, gotten from the City_Populations method}

\item{pct_primary, pct_secondary}{percentage of the county that is a primary and secondary shopper}
}
\value{
sets values for the number of primary and secondary shoppers
}
\description{
Calculation for a quarter circle.
}
