% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primary_Shopper_Count.R
\name{Primary_Shopper_Count}
\alias{Primary_Shopper_Count}
\title{Calculates the number of primary shoppers in each marker}
\usage{
Primary_Shopper_Count(
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list
)
}
\arguments{
\item{pct_metro_prim, pct_rural_prim, pct_town_prim}{The percentage of
primary shoppers in metro,rural and town markets default taken as
50\%,30\%,30\% respectively unless specified otherwise.}

\item{county_pop, towns_pop}{population of the county and population of
all towns in the county}

\item{pct_county}{percentage of county in our market}

\item{metro_list, town_list}{lists of population of store location and
populations of towns in the neighborhood}
}
\value{
outputs the total number of primary shoppers for the store
}
\description{
Calculates the number of primary shoppers in each marker
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6. This function calls Rural_Population() and
City_Populations()
}
\examples{
Primary_Shopper_Count(county_pop=18000,towns_pop=9786,pct_county=17.7,
town_list=list(77),metro_list=list(2650))

}
