% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pull_Cities.R
\name{Pull_Cities}
\alias{Pull_Cities}
\title{Takes a Vector of States and Retrieves All Cities in State(s)}
\usage{
Pull_Cities(df_grocery_all, state_list)
}
\arguments{
\item{df_grocery_all}{A data frame containing all the store locations pulled from
the googleway::google_places API.}

\item{state_list}{A list of unique state values to iterate through.}
}
\value{
returns a data frame of all the cities in a state(s) called all_cities
}
\description{
the function takes a list of states and passes them through a loop
to pull all cities in the provided states.
}
\details{
values in the list are filtered using unique() to avoid repetitive loops.
}
\author{
Harun Celik
}
