% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pull_Counties.R
\name{Pull_Counties}
\alias{Pull_Counties}
\title{Takes a Vector of States and Retrieves All Counties in State(s)}
\usage{
Pull_Counties(
  df_grocery_all = df_grocery_all,
  state_list = unique(df_grocery_all$state)
)
}
\arguments{
\item{df_grocery_all}{A data frame containing all the store locations pulled from
the googleway::google_places API.}

\item{state_list}{A list of unique state values to iterate through.}
}
\value{
returns a data frame of all the counties in a state(s) called all_counties
}
\description{
the function takes a list of states and passes them through a loop
to pull all counties in the provided states.
}
\details{
values in the list are filtered using unique() to avoid repetitive loops.
}
\author{
Harun Celik
}
