% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadrant_Calculator.R
\name{Quadrant_Calculator}
\alias{Quadrant_Calculator}
\title{Determines Distances to Stores in Quadrant Areas}
\usage{
Quadrant_Calculator(df_grocery_only, df_geocode)
}
\arguments{
\item{df_grocery_only}{a data frame containing only grocery store observations.
Returned from a call to the googleway::google_places api.}

\item{df_geocode}{a data frame containing only the lat and lng values of a geocoded
address. The geocode is returned from googleway::google_geocode.}
}
\value{
this function returns a list of the four distances for the closest stores
in each quadrant and the data frame 'df_circle_buffer' to be inherited by
DSPGrocery::Create_Circle_Buffer.
}
\description{
Finds the distance to the nearest store in each quadrant.
No return, just sets four global variables
}
\author{
Alex Cory
}
