% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadrant_Area.R
\name{QuadrantArea}
\alias{QuadrantArea}
\title{Calculation for a quarter circle.}
\usage{
QuadrantArea(user_area)
}
\arguments{
\item{user_area}{user defined area circle}
}
\value{
outputs output_area
}
\description{
Calculation for a quarter circle.
}
\examples{
quadrant_area(user_area = 4)
}
