% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rare_Shopper_Count.R
\name{Rare_Shopper_Count}
\alias{Rare_Shopper_Count}
\title{Calculates the Number of Rare Shoppers in Each Marker}
\usage{
Rare_Shopper_Count(
  pct_metro_rare = 10,
  pct_rural_rare = 20,
  pct_town_rare = 20,
  metro_pop,
  town_pop,
  rural_pop
)
}
\arguments{
\item{pct_metro_rare}{The percentage of secondary shoppers in metro market
default taken as 10\% unless specified otherwise.}

\item{pct_rural_rare}{The percentage of secondary shoppers in rural market
default taken as 20\% unless specified otherwise.}

\item{pct_town_rare}{The percentage of secondary shoppers in town market
default taken as 20\% unless specified otherwise.}

\item{metro_pop}{Population of the store location}

\item{town_pop}{Population of all towns in the county}

\item{rural_pop}{Population of rural cities in our market}
}
\value{
Outputs the total number of Rare shoppers for the store
}
\description{
This function calculates the number of rare shoppers given the
population in each market type.
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6.
}
\examples{
#Rare_Shopper_Count()

}
\seealso{
\code{\link[=Primary_Shopper_Count]{Primary_Shopper_Count()}}

\code{\link[=Secondary_Shopper_Count]{Secondary_Shopper_Count()}}
}
\author{
Srika Raja
}
