% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rural_Pop.R
\name{Rural_Pop}
\alias{Rural_Pop}
\title{Finds the Population of everyone who is in the Area but not in a Town}
\usage{
Rural_Pop(df_city_pop, geo_county, df_grocery_only, df_geocode, popbinder)
}
\arguments{
\item{df_city_pop}{df with city populations}

\item{geo_county}{name if county}

\item{df_grocery_only}{API call with just grocery stores}

\item{df_geocode}{latlong of store}

\item{popbinder}{inherited list from the call to DSPGrocery::Calc_Market_Size
containing df_city_pop and county_name.}
}
\value{
Count of people who would shop at the store who do not live in a city
}
\description{
Determines population that does not live in a town or a city but would likely
shop at the proposed store
}
\author{
Alex Cory
}
