% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rural_Population.R
\name{Rural_Population}
\alias{Rural_Population}
\title{Calculation for a quarter circle.}
\usage{
Rural_Population(county_pop, towns_pop, pct_county)
}
\arguments{
\item{county_pop}{the population of the whole county}

\item{towns_pop}{the sum of all the towns populations in the county}

\item{pct_county}{the percent of the county in the area}
}
\value{
outputs the rural population
}
\description{
Calculation for a quarter circle.
}
