% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rural_Shoppers_Count.R
\name{Rural_Shoppers_Count}
\alias{Rural_Shoppers_Count}
\title{Calculates Count of Rare Shoppers for a Rural Market}
\usage{
Rural_Shoppers_Count(pct_primary = 30, pct_secondary = 50, pct_rare = 20)
}
\arguments{
\item{pct_primary}{percentage of the county that is a primary shopper defaulted
to 30\%.}

\item{pct_secondary}{percentage of the county that is a secondary shopper defaulted
to 50\%.}

\item{pct_rare}{percentage of the county that is a rare shopper defaulted
to 20\%.}
}
\value{
number of shoppers rural shoppers based on the number of primary,
secondary, and rare shoppers
}
\description{
Calculates Count of Rare Shoppers for a Rural Market
}
