% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secondary_Shopper_Count.R
\name{Secondary_Shopper_Count}
\alias{Secondary_Shopper_Count}
\title{Calculates the Number of Secondary Shoppers in Each Market}
\usage{
Secondary_Shopper_Count(
  pct_metro_sec = 40,
  pct_rural_sec = 50,
  pct_town_sec = 50,
  metro_pop,
  town_pop,
  rural_pop
)
}
\arguments{
\item{pct_metro_sec}{The percentage of secondary shoppers in metro market
default taken as 40\% unless specified otherwise.}

\item{pct_rural_sec}{The percentage of secondary shoppers in rural market
default taken as 50\% unless specified otherwise.}

\item{pct_town_sec}{The percentage of secondary shoppers in town market
default taken as 50\% unless specified otherwise.}

\item{metro_pop}{Population of the store location}

\item{town_pop}{Population of all towns in the county}

\item{rural_pop}{Population of rural cities in our market}
}
\value{
Outputs the total number of secondary shoppers for the store
}
\description{
This function calculates the number of secondary shoppers given the
population in each market type.
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6.
}
\examples{
Secondary_Shopper_Count()

}
\seealso{
\code{\link[=Primary_Shopper_Count(),Rare_Shopper_Count]{Primary_Shopper_Count(),Rare_Shopper_Count()}}
}
\author{
Srika Raja
}
