% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State_Adj_Capita_Grocery_Spend.R
\name{State_Adj_Capita_Grocery_Spend}
\alias{State_Adj_Capita_Grocery_Spend}
\title{Calculates the average per capita spend adjusting the estimated price increase and state index}
\usage{
State_Adj_Capita_Grocery_Spend(state_index)
}
\arguments{
\item{state_index}{the ratio of the current price of the basket to the price of the basket during the base year}
}
\value{
outputs the average per capita grocery spend after adjusting for the expected price increase and state index
}
\description{
Calculates the average per capita spend adjusting the estimated price increase and state index
}
\details{
This function is taken from the Estimating_Revenue.xlsx and calculated in Step two. This function calls \code{Adj_Capita_Grocery_Spend()}.
}
\examples{
State_Adj_Capita_Grocery_Spend(99)

}
