% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Primary_Shoppers.R
\name{Total_Spend_Primary_Shoppers}
\alias{Total_Spend_Primary_Shoppers}
\title{Calculates the money spent by primary shoppers}
\usage{
Total_Spend_Primary_Shoppers(
  per_of_grocery_spend_prim = 60,
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list,
  state_index,
  est_per_price_increase,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_prim}{the percentage of groceries primary
shopper will spend at the proposed store}

\item{pct_metro_prim, pct_rural_prim, pct_town_prim}{The percentage of
primary shoppers in metro,rural and town markets default taken as
50\%,30\%,30\% respectively unless specified otherwise.}

\item{county_pop, towns_pop}{population of the county and population of
all towns in the county}

\item{pct_county}{percentage of county in our market}

\item{metro_list, town_list}{lists of population of store location and
populations of towns in the neighborhood}
}
\value{
outputs the total spending by Primary Shoppers at the proposed store
}
\description{
Calculates the money spent by primary shoppers
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Primary_Shopper_Count()
}
