% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Secondary_Shoppers.R
\name{Total_Spend_Secondary_Shoppers}
\alias{Total_Spend_Secondary_Shoppers}
\title{Calculates the Money Spent by Secondary Shoppers}
\usage{
Total_Spend_Secondary_Shoppers(
  per_of_grocery_spend_sec = 25,
  pct_metro_sec = 40,
  pct_rural_sec = 50,
  pct_town_sec = 50,
  metro_pop,
  town_pop,
  rural_pop,
  state_index,
  est_per_price_increase = 7,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_sec}{the percentage of groceries secondary
shoppers will spend at the proposed store}

\item{pct_metro_sec}{The percentage of secondary shoppers in metro market
default taken as 40\% unless specified otherwise.}

\item{pct_rural_sec}{The percentage of secondary shoppers in rural market
default taken as 50\% unless specified otherwise.}

\item{pct_town_sec}{The percentage of secondary shoppers in town market
default taken as 50\% unless specified otherwise.}

\item{metro_pop}{Population of the store location}

\item{town_pop}{Population of all towns in the county}

\item{rural_pop}{Population of rural cities in our market}

\item{state_index}{the ratio of the current price of the basket to the price of
the basket during the base year}

\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent
default taken as 7 for 2023.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
Outputs the total spending by secondary shoppers at the proposed store
}
\description{
This function returns the total money spent by secondary shoppers visiting a store.
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Secondary_Shopper_Count()
}
\examples{
Total_Spend_Secondary_Shoppers(state_index=99)

}
\seealso{
\code{\link[=Total_Spend_Primary_Shoppers(),Total_Spend_Rare_Shoppers]{Total_Spend_Primary_Shoppers(),Total_Spend_Rare_Shoppers()}}
}
\author{
Srika Raja
}
