% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Area.R
\name{TotalArea}
\alias{TotalArea}
\title{Calculation for a quarter circle.}
\usage{
TotalArea(north_quad, east_quad, south_quad, west_quad)
}
\arguments{
\item{north_quad, east_quad, south_quad, west_quad}{user defined quadrant sizes}
}
\value{
outputs sum of quadrant areas
}
\description{
Calculation for a quarter circle.
}
