% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_Zoner.R
\name{UTM_Zoner}
\alias{UTM_Zoner}
\title{Finding UTM Zone from Longitude in North America}
\usage{
UTM_Zoner(lng)
}
\arguments{
\item{lng}{longitude}
}
\value{
zone the UTM zone
}
\description{
Allows the user to enter a longitude and have it return the
UTM zone. Works only in North America.
}
\author{
Alex Cory
}
