% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_Capita_Grocery_Spend.R
\name{Adj_Capita_Grocery_Spend}
\alias{Adj_Capita_Grocery_Spend}
\title{Calculates the average per capita spend adjusting the estimated price increase}
\usage{
Adj_Capita_Grocery_Spend(est_per_price_increase)
}
\arguments{
\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent.}
}
\value{
Outputs the average per capita grocery spend after adjusting for the expected price increase.
}
\description{
Calculates the average per capita spend adjusting the estimated price increase
}
\details{
This function is taken from the Estimating_Revenue.xlsx and calculated in Step two. This function calls \code{Avg_Capita_Grocery_Spending()}.
}
\examples{
Adj_Capita_Grocery_Spend(7)
}
