% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adj_Capita_Grocery_Spend.R
\name{Adj_Capita_Grocery_Spend}
\alias{Adj_Capita_Grocery_Spend}
\title{Calculates the average per capita spend adjusting the estimated price increase}
\usage{
Adj_Capita_Grocery_Spend(
  est_per_price_increase,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
Outputs the average per capita grocery spend after adjusting for the expected price increase.
}
\description{
Calculates the average per capita spend adjusting the estimated price increase
}
\details{
This function is taken from the Estimating_Revenue.xlsx and calculated in Step two. This function calls \code{Avg_Capita_Grocery_Spending()}.
}
\examples{
Adj_Capita_Grocery_Spend(7)
}
