% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cost_of_Goods_Sold.R
\name{Cost_of_Goods_Sold}
\alias{Cost_of_Goods_Sold}
\title{Calculation of the Cost of Goods Sold}
\usage{
Cost_of_Goods_Sold(Total_Estimated_Revenue)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of a total estimated revenue calculation}
}
\value{
Outputs cost of goods sold (in dollars)
}
\description{
Calculation of the Cost of Goods Sold
}
\details{
Calls Gross_Margin() function to provide the total subtracted from revenue to
calculate cost of goods
This function was taken from Estimating_Expense.xlxs (Step 6)
}
\examples{
# Cost_of_Goods_Sold(Total_Estimated_Revenue = 2000000)

}
