% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cost_of_Goods_Sold.R
\name{Cost_of_Goods_Sold}
\alias{Cost_of_Goods_Sold}
\title{Calculates the Cost of Goods Sold Based on Gross Margin for Grocery Stores}
\usage{
Cost_of_Goods_Sold(Total_Estimated_Revenue, Gross_Margin_Percentage = 0.2446)
}
\arguments{
\item{Total_Estimated_Revenue}{Total estimated revenue based on market size/location.}

\item{Gross_Margin_Percentage}{Percentage of total estimated revenue expected as gross margin
selected via user input (default from Bizminer).}
}
\value{
The output returns the estimated dollar amount for the cost of goods sold for
a hypothetical grocery store in a given location.
}
\description{
This function calculates the estimated cost of goods sold based on
the selected gross margin percentage for a hypothetical grocery store
in a user-selected location.
}
\details{
This function calls the "Gross_Margin()" function from the DSPG package in its body.
This function is based upon calculations originally
formulated by FFED ISU Extension and Outreach.
}
\author{
Aaron Null
}
