% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cost_of_Goods.R
\name{Cost_of_Goods_Sold}
\alias{Cost_of_Goods_Sold}
\title{Calculation of the Cost of Goods Sold}
\usage{
Cost_of_Goods_Sold(Total_Estimated_Revenue, Percentage = 0.7816)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of a total estimated revenue calculation}

\item{Percentage}{Percentage of total estimated revenue spent on supply of goods. Default value is .7816 taken from BizMiner (3 year average).}
}
\value{
Outputs cost of goods sold (in dollars)
}
\description{
Calculation of the Cost of Goods Sold
}
\examples{
# Cost_of_Goods_Sold(Total_Estimated_Revenue = 2000000, Percentage = .7816)

}
