% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depreciation_1.R
\name{Depreciation_1}
\alias{Depreciation_1}
\title{Calculation of Depreciation (for Building Ownership)}
\usage{
Depreciation_1(
  Building_Remodeling,
  Parking_Lot_Improvements,
  Shelving_Check_Out_Counters,
  Computer_Equipment_POS,
  Vehicles,
  Display_Cases,
  Refrigeration,
  Freezers,
  Meat_Cutting_Equipment,
  Miscellaneous_Assets_1,
  Miscellaneous_Assets_1_Use_Life,
  Miscellaneous_Assets_2,
  Miscellaneous_Assets_2_Use_Life,
  Miscellaneous_Assets_3,
  Miscellaneous_Assets_3_Use_Life
)
}
\arguments{
\item{Building_Remodeling}{Cost  of Building/Remodeling}

\item{Parking_Lot_Improvements}{Depreciation of Parking Lot Improvements}

\item{Shelving_Check_Out_Counters}{Depreciation of Shelving/Check Out Counters}

\item{Computer_Equipment_POS}{Cost of Computer System/POS}

\item{Vehicles}{Cost of Vehicles}

\item{Display_Cases}{Cost of Display cases}

\item{Refrigeration}{Cost of Refrigeration}

\item{Freezers}{Cost of Freezers}

\item{Meat_Cutting_Equipment}{Cost of Meat-cutting Equipment}

\item{Miscellaneous_Assets_1}{(Optional) Cost of Miscellaneous Asset (1)}

\item{Miscellaneous_Assets_1_Use_Life}{(Optional) Use Life of Miscellaneous Asset (1)}

\item{Miscellaneous_Assets_2}{(Optional) Cost of Miscellaneous Asset (2)}

\item{Miscellaneous_Assets_2_Use_Life}{(Optional) Use Life of Miscellaneous Asset (2)}

\item{Miscellaneous_Assets_3}{(Optional) Cost of Miscellaneous Asset (3)}

\item{Miscellaneous_Assets_3_Use_Life}{(Optional) Use life of Miscellaneous Asset (3)}
}
\value{
Outputs total estimated loss from depreciation for property owners
}
\description{
Calculation of Depreciation (for Building Ownership)
}
\details{
This function is taken from the "Estimating Expense" excel sheet (step two)
}
