% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depreciation_1.R
\name{Depreciation_1}
\alias{Depreciation_1}
\title{Calculates the Cost of Depreciation Based on List of Assets (For a Building Owner)}
\usage{
Depreciation_1(
  Building_Remodeling,
  Parking_Lot_Improvements,
  Shelving_Check_Out_Counters,
  Computer_Equipment_POS,
  Vehicles,
  Display_Cases,
  Refrigeration,
  Freezers,
  Meat_Cutting_Equipment,
  Miscellaneous_Assets_1,
  Miscellaneous_Assets_1_Use_Life,
  Miscellaneous_Assets_2,
  Miscellaneous_Assets_2_Use_Life,
  Miscellaneous_Assets_3,
  Miscellaneous_Assets_3_Use_Life
)
}
\arguments{
\item{Building_Remodeling}{Cost  of Building/Remodeling.}

\item{Parking_Lot_Improvements}{Cost of Parking Lot Improvements.}

\item{Shelving_Check_Out_Counters}{Cost of Shelving/Check Out Counters.}

\item{Computer_Equipment_POS}{Cost of Computer System/POS.}

\item{Vehicles}{Cost of Vehicles.}

\item{Display_Cases}{Cost of Display cases.}

\item{Refrigeration}{Cost of Refrigeration.}

\item{Freezers}{Cost of Freezers.}

\item{Meat_Cutting_Equipment}{Cost of Meat-cutting Equipment.}

\item{Miscellaneous_Assets_1}{(Optional) Cost of Miscellaneous Asset (1)}

\item{Miscellaneous_Assets_1_Use_Life}{(Optional) Use Life of Miscellaneous Asset (1)}

\item{Miscellaneous_Assets_2}{(Optional) Cost of Miscellaneous Asset (2)}

\item{Miscellaneous_Assets_2_Use_Life}{(Optional) Use Life of Miscellaneous Asset (2)}

\item{Miscellaneous_Assets_3}{(Optional) Cost of Miscellaneous Asset (3)}

\item{Miscellaneous_Assets_3_Use_Life}{(Optional) Use life of Miscellaneous Asset (3)}
}
\value{
The output returns the estimated dollar amount lost annually to asset
depreciation for a hypoththetical grocery store.
}
\description{
This function calculates the estimated cost of the annual depreciation of assets
for grocery stores (specifically for scenario 1 in which the building
is owned, not rented).
}
\details{
This function is based upon calculations originally
formulated by FFED ISU Extension and Outreach. The values of the various assets
are divided by their use life to get the annual loss from depreciation for each asset.
These values are added up to get the total annual loss from asset
depreciation/amortization. Users can enter in up to 3 miscellaneous assets
and each of their corresponding use-lives.
}
\examples{
Depreciation_1(
  Building_Remodeling = 900000,
  Parking_Lot_Improvements = 50000,
  Shelving_Check_Out_Counters = 60000,
  Computer_Equipment_POS = 15000,
  Vehicles = 12000,
  Display_Cases = 15000,
  Refrigeration = 20000,
  Freezers = 20000,
  Meat_Cutting_Equipment = 10000,
  Miscellaneous_Assets_1 = 10000,
  Miscellaneous_Assets_1_Use_Life = 10,
  Miscellaneous_Assets_2 = 10000,
  Miscellaneous_Assets_2_Use_Life = 10,
  Miscellaneous_Assets_3 = 10000,
  Miscellaneous_Assets_3_Use_Life = 10)


}
\author{
Aaron Null
}
