% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Depreciation_2.R
\name{Depreciation_2}
\alias{Depreciation_2}
\title{Calculation of Depreciation (for Leasing)}
\usage{
Depreciation_2(
  Leasehold_Improvements,
  Leasehold_Improvements_Use_Life,
  Shelving_Check_Out_Counters,
  Computer_Equipment_POS,
  Vehicles,
  Display_Cases,
  Refrigeration,
  Freezers,
  Meat_Cutting_Equipment,
  Miscellaneous_Assets_1,
  Miscellaneous_Assets_1_Use_Life,
  Miscellaneous_Assets_2,
  Miscellaneous_Assets_2_Use_Life,
  Miscellaneous_Assets_3,
  Miscellaneous_Assets_3_Use_Life
)
}
\value{
Outputs total estimated loss from depreciation for leasers
}
\description{
Calculation of Depreciation (for Leasing)
}
\details{
This function is taken from the "Estimating Expense" excel sheet (step three)
}
\examples{
#

}
