% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Employee_Wages.R
\name{Employee_Wages}
\alias{Employee_Wages}
\title{Calculates the Estimated Cost of Employee Wages for Grocery Stores}
\usage{
Employee_Wages(Total_Estimated_Revenue, Percentage = 0.0942)
}
\arguments{
\item{Total_Estimated_Revenue}{Total estimated revenue based on market size/location.}

\item{Percentage}{Percentage of total estimated revenue spent on employee wages selected
via user input (default from Bizminer).}
}
\value{
The output returns the estimated dollar amount spent on employee wages for
a hypothetical grocery store in a given location.
}
\description{
This function calculates the estimated cost of employee wages for a hypothetical store in
a user-selected location based off of a user-selected percentage parameter. If not
specified by the user, this percentage defaults to the industry average
provided by Bizminer.
}
\details{
This function employs a default percentage value from the financial analysis
and market research firm Bizminer and is based upon calculations originally
formulated by FFED ISU Extension and Outreach.
}
\author{
Aaron Null
}
