% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Employee_Wages.R
\name{Employee_Wages}
\alias{Employee_Wages}
\title{Calculation of Employee Wages}
\usage{
Employee_Wages(Total_Estimated_Revenue, Percentage = 0.0942)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of total estimated revenue calculation sheet}

\item{Percentage}{Percentage of revenue spent on employee wages (default taken from Bizminer)}
}
\value{
Outputs dollars spent on employee wages (based off of provided percentage)
}
\description{
Calculation of Employee Wages
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and
calculated in Step Six.
}
\examples{
Employee_Wages(Total_Estimated_Revenue = 2000000, Percentage = .0942)

}
