% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gross_Margin.R
\name{Gross_Margin}
\alias{Gross_Margin}
\title{Calculation of Gross Margin}
\usage{
Gross_Margin(Total_Estimated_Revenue, Percentage = 0.2446)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of a total estimated revenue calculation}

\item{Percentage}{Percentage of Revenue as Gross Margin (Revenue - Cost of Goods)}
}
\value{
Outputs Gross Margin (in dollars) depending on the percentage provided
(default from Bizminer)
}
\description{
Calculation of Gross Margin
}
\details{
This function is taken from the "Estimating Expense" excel sheet (step six)
}
\examples{
# Gross_Margin(Total_Estimated_Revenue = 2000000, Percentage = .2446)

}
