% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gross_Margin.R
\name{Gross_Margin}
\alias{Gross_Margin}
\title{Calculation of Gross Margin}
\usage{
Gross_Margin(Total_Estimated_Revenue)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of a total estimated revenue calculation}
}
\value{
Outputs Gross Margin (Total Estimated Revenue - Cost of Goods Sold) (in dollars)
}
\description{
Calculation of Gross Margin
}
\details{
This function is taken from the "Estimating Expense" excel sheet
}
\examples{
# Gross_Margin(Total_Estimated_Revenue = 2000000)

}
