% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interest_Expense.R
\name{Interest_Expense}
\alias{Interest_Expense}
\title{Calculates the Annual Cost of Interest on a Loan}
\usage{
Interest_Expense(Loan_Amount, Interest_Rate)
}
\arguments{
\item{Loan_Amount}{Dollar amount of the user's loan.}

\item{Interest_Rate}{The annual interest rate on that loan.}
}
\value{
The output returns the estimated dollar amount spent on interest for a loan.
}
\description{
This function calculates the annual cost of interest on a loan. The amount of the loan
and the annual interest rate is provided by the user.
}
\details{
This function calculates the dollar amount spent on interest for a loan by multiplying
the loan by the annual interest rate (provided by the user).
}
\author{
Aaron Null
}
