% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interest_Income.R
\name{Interest_Income}
\alias{Interest_Income}
\title{Calculation of Interest Income}
\usage{
Interest_Income(Total_Estimated_Revenue, Percentage = 4e-04)
}
\arguments{
\item{Total_Estimated_Revenue}{Calculation of Total_Estimated Revenue}

\item{Percentage}{Percentage of revenue of the estimated income from interest}
}
\value{
Outputs first-year interest income (in dollars)
}
\description{
Calculation of Interest Income
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and
calculated in Step Eight.
}
\examples{
Interest_Income(Total_Estimated_Revenue = 2000000, Percentage = .0004)

}
