% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Officer_Compensation.R
\name{Officer_Compensation}
\alias{Officer_Compensation}
\title{Calculation of Officer Compensation}
\usage{
Officer_Compensation(Total_Estimated_Revenue, Percentage = 0.0118)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of total estimated revenue calculation sheet}

\item{Percentage}{Percentage of revenue spent on officer compensation (default from Bizminer)}
}
\value{
Outputs dollars spent on officer compensation (Total Estimated Revenue * Percentage) (in dollars)
}
\description{
Calculation of Officer Compensation
}
\details{
This function was taken from the "Estimating_Expenses.xlsx" sheet in Step 6.
}
\examples{
# Officer_Compensation(Total_Estimated_Revenue = 2000000, Percentage = .0118)

}
