% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operating_Income_Loss.R
\name{Operating_Income_Loss}
\alias{Operating_Income_Loss}
\title{Calculation of Revenue Surplus after the Subtraction of Cost of Goods and other Operating
Costs (Interim Result - Not Final)}
\usage{
Operating_Income_Loss(
  Total_Estimated_Revenue,
  Gross_Margin_Percentage = 0.2446
)
}
\arguments{
\item{Total_Estimated_Revenue}{Total estimated revenue based on market size/location.}

\item{Gross_Margin_Percentage}{User-selected gross margin percentage
(default taken from Bizminer).}
}
\value{
The output returns the estimated revenue surplus (revenue left over) after the
subtraction of the cost of goods sold and operating expenses (minus depreciation).
}
\description{
This function calculates the estimated revenue surplus after subtracting the cost of goods,
officer compensation, employee wages and other operating expenses and outputs an interim
profit figure before depreciation, interest income, interest expense and secondary sources
of income are taken into account.
}
\details{
This function employs the user-selected gross margin percentage as an argument and calls
the functions "Cost_of_Goods_Sold()", "Officer_Compensation()", "Employee_Wages()" and
"Other_Operating_Expense()". This function is based upon calculations originally
formulated by FFED ISU Extension and Outreach.
}
\author{
Aaron Null
}
