% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Income.R
\name{Other_Income}
\alias{Other_Income}
\title{Calculation of Other Income}
\usage{
Other_Income(Total_Estimated_Revenue)
}
\arguments{
\item{Total_Estimated_Revenue}{Calculation of Total_Estimated Revenue}
}
\value{
Outputs "Other Income" (in dollars)
}
\description{
Calculation of Other Income
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and calculated in Step Eight.
Percentages are taken from Bizminer.
}
\examples{
# Other_Income(Total_Estimated_Revenue = 2000000)

}
