% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_Operating_Expense.R
\name{Other_Operating_Expense}
\alias{Other_Operating_Expense}
\title{Calculation of other operating expenses (miscellaneous)}
\usage{
Other_Operating_Expense(Total_Estimated_Revenue)
}
\arguments{
\item{Total_Estimated_Revenue}{Output of total estimated revenue calculation sheet}

\item{Percentage}{Percentage of total estimated revenue spent on other operating expenses.
Percentages are taken from Bizminer.}
}
\value{
Outputs other operating expenses (in dollars)
}
\description{
Calculation of other operating expenses (miscellaneous)
}
\details{
This function is taken from the "Estimating_Expenses.xlsx" sheet and calculated in Step Six)
}
\examples{
# Employee_Wages(Total_Estimated_Revenue = 2000000)

}
