% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pct_County.R
\name{Pct_County}
\alias{Pct_County}
\title{Calculates Sum of Quadrant Areas.}
\usage{
Pct_County(north_quad, east_quad, south_quad, west_quad, county_size)
}
\arguments{
\item{north_quad}{user defined size of north quadrant}

\item{east_quad}{user defined size of east quadrant}

\item{south_quad}{user defined size of south quadrant}

\item{west_quad}{user defined size of west quadrant}

\item{county_size}{User defined Size of county}
}
\value{
outputs sum of quadrant areas
}
\description{
Calculates Sum of Quadrant Areas.
}
