% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pct_County.R
\name{Pct_County}
\alias{Pct_County}
\title{Calculation for a quarter circle.}
\usage{
Pct_County(north_quad, east_quad, south_quad, west_quad, county_size)
}
\arguments{
\item{north_quad, east_quad, south_quad, west_quad}{user defined quadrant sizes}
}
\value{
outputs sum of quadrant areas
}
\description{
Calculation for a quarter circle.
}
