% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pop_Binder.R
\name{Pop_Binder}
\alias{Pop_Binder}
\title{Creating a Dataframe from an Address}
\usage{
Pop_Binder(address)
}
\arguments{
\item{address}{the address must be in the standard Google maps form}
}
\description{
Function that takes in an address and returns a dataframe with cities in the
county and their populations in a tidy format.
}
\details{
Makes a TidyCensus and Google Places API call, needs key
}
\examples{
# Pop_Binder("223 S Main St, Lake View, IA, Unites States, Iowa")
}
\author{
Alex Cory
}
