% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primary_Shopper_Count.R
\name{Primary_Shopper_Count}
\alias{Primary_Shopper_Count}
\title{Calculates the Number of Primary Shoppers in Each Marker}
\usage{
Primary_Shopper_Count(
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list
)
}
\arguments{
\item{pct_metro_prim}{The percentage of primary shoppers in metro market
default taken as 50\% unless specified otherwise.}

\item{pct_rural_prim}{The percentage of primary shoppers in rural market
default taken as 30\% unless specified otherwise.}

\item{pct_town_prim}{The percentage of primary shoppers in town market
default taken as 30\% unless specified otherwise.}

\item{county_pop}{Population of the county.}

\item{towns_pop}{Population of all towns in the county.}

\item{pct_county}{Percentage of county in our market.}

\item{metro_list}{Lists of population of store location.}

\item{town_list}{List of population of towns in the neighborhood.}
}
\value{
Outputs the total number of primary shoppers for the store
}
\description{
This function calculates the number of primary shoppers given the
population in each market type.
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6. This function calls Rural_Population() and
City_Populations()
}
\examples{
Primary_Shopper_Count(county_pop=18000,
                     towns_pop=9786,
                     pct_county=17.7,
                     town_list=list(77),
                     metro_list=list(2650))

}
\author{
Srika Raja
}
