% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Primary_Shopper_Count.R
\name{Primary_Shopper_Count}
\alias{Primary_Shopper_Count}
\title{Calculates the Number of Primary Shoppers in Each Marker}
\usage{
Primary_Shopper_Count(
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  metro_pop,
  town_pop,
  rural_pop
)
}
\arguments{
\item{pct_metro_prim}{The percentage of primary shoppers in metro market
default taken as 50\% unless specified otherwise.}

\item{pct_rural_prim}{The percentage of primary shoppers in rural market
default taken as 30\% unless specified otherwise.}

\item{pct_town_prim}{The percentage of primary shoppers in town market
default taken as 30\% unless specified otherwise.}

\item{metro_pop}{Population of the store location}

\item{town_pop}{Population of all towns in the county}

\item{rural_pop}{Population of rural cities in our market}
}
\value{
Outputs the total number of primary shoppers for the store
}
\description{
This function calculates the number of primary shoppers given the
population in each market type.
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6.
}
\examples{
#Primary_Shopper_Count()

}
\seealso{
\code{\link[=Secondary_Shopper_Count]{Secondary_Shopper_Count()}}

\code{\link[=Rare_Shopper_Count]{Rare_Shopper_Count()}}
}
\author{
Srika Raja
}
