% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrant_area.R
\name{quadrant_area}
\alias{quadrant_area}
\title{Calculation for a quarter circle.}
\usage{
quadrant_area(user_area)
}
\arguments{
\item{user_area}{user defined area circle}
}
\value{
outputs output_area
}
\description{
Calculation for a quarter circle.
}
\examples{
quadrant_area(user_area = 4)
}
