% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadrant_Area.R
\name{Quadrant_Area}
\alias{Quadrant_Area}
\title{Calculation for a quarter circle.}
\usage{
Quadrant_Area(user_area)
}
\arguments{
\item{user_area}{user defined area circle}
}
\value{
outputs output_area
}
\description{
Calculation for a quarter circle.
}
\examples{
Quadrant_Area(user_area = 4)
}
