% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quadrant_Area.R
\name{Quadrant_Area}
\alias{Quadrant_Area}
\title{Calculates the Area for a  Quarter Circle.}
\usage{
Quadrant_Area(user_area)
}
\arguments{
\item{user_area}{user defined area circle}
}
\value{
outputs output_area
}
\description{
Calculates the Area for a  Quarter Circle.
}
\examples{
Quadrant_Area(user_area = 4)

}
