% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rare_Shopper_Count.R
\name{Rare_Shopper_Count}
\alias{Rare_Shopper_Count}
\title{Calculates the number of Rare shoppers in each marker}
\usage{
Rare_Shopper_Count(
  pct_metro_rare = 10,
  pct_rural_rare = 20,
  pct_town_rare = 20,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list
)
}
\arguments{
\item{county_pop, towns_pop}{population of the county and population of
all towns in the county}

\item{pct_county}{percentage of county in our market}

\item{metro_list, town_list}{lists of population of store location and
populations of towns in the neighborhood}

\item{pct_metro_sec, pct_rural_sec, pct_town_sec}{The percentage of
Rare shoppers in metro,rural and town markets default taken as
10\%,20\%,20\% respectively unless specified otherwise.}
}
\value{
outputs the total number of Rare shoppers for the store
}
\description{
Calculates the number of Rare shoppers in each marker
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6. This function calls Rural_Population() and
City_Populations()
}
\examples{
Rare_Shopper_Count(county_pop=18000,towns_pop=9786,
pct_county=17.7,town_list=list(77),metro_list=list(2650))

}
