% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rare_Shopper_Count.R
\name{Rare_Shopper_Count}
\alias{Rare_Shopper_Count}
\title{Calculates the number of Rare shoppers in each marker}
\usage{
Rare_Shopper_Count(
  pct_metro_rare = 10,
  pct_rural_rare = 20,
  pct_town_rare = 20,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list
)
}
\arguments{
\item{pct_metro_rare}{The percentage of secondary shoppers in metro market
default taken as 10\% unless specified otherwise.}

\item{pct_rural_rare}{The percentage of secondary shoppers in rural market
default taken as 20\% unless specified otherwise.}

\item{pct_town_rare}{The percentage of secondary shoppers in town market
default taken as 20\% unless specified otherwise.}

\item{county_pop}{population of the county.}

\item{towns_pop}{population of all towns in the county.}

\item{pct_county}{percentage of county in our market.}

\item{metro_list}{lists of population of store location.}

\item{town_list}{list of population of towns in the neighborhood.}
}
\value{
outputs the total number of Rare shoppers for the store
}
\description{
Calculates the number of Rare shoppers in each marker
}
\details{
This function is taken from the estimating Market Size.xlsx
and calculated in step 6. This function calls Rural_Population() and
City_Populations()
}
\examples{
Rare_Shopper_Count(county_pop=18000,towns_pop=9786,pct_county=17.7,town_list=list(77),metro_list=list(2650))

}
