% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rural_Shoppers_Count.R
\name{Rural_Shoppers_Count}
\alias{Rural_Shoppers_Count}
\title{Calculation for a shoppers from Primary secondary and rare categories in a Rural Market}
\usage{
Rural_Shoppers_Count(pct_primary = 30, pct_secondary = 50, pct_rare = 20)
}
\arguments{
\item{pct_primary, pct_secondary, pct_rare}{percentage of the county that is a primary,
secondary and rare shopper default taken as 30,50,20\% unless specified otherwise}

\item{pop}{population in rural market, gotten from the Rural_Populations method}
}
\value{
sets values for the number of primary, secondary, and rare shoppers
}
\description{
Calculation for a shoppers from Primary secondary and rare categories in a Rural Market
}
