% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State_Adj_Capita_Grocery_Spend.R
\name{State_Adj_Capita_Grocery_Spend}
\alias{State_Adj_Capita_Grocery_Spend}
\title{Calculates Average per Capita Spending Adjusted by Estimated Price Increase and State Index}
\usage{
State_Adj_Capita_Grocery_Spend(
  state_index,
  est_per_price_increase = 7,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{state_index}{the ratio of the current price of the basket to the price of the
basket during the base year}

\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent
default taken as 7 for 2023.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
Outputs the average per capita grocery spend after adjusting for the expected
price increase and state index
}
\description{
This function returns the average per capita spending after adjusting for
percentage price increase and rural price parities from the base year(2022).
}
\details{
This function is taken from the Estimating_Revenue.xlsx and calculated in Step three.
This function calls \code{Adj_Capita_Grocery_Spend()}.
}
\examples{
State_Adj_Capita_Grocery_Spend(99,7)

}
\author{
Srika Raja
}
