% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Estimate_Revenue.R
\name{Total_Estimate_Revenue}
\alias{Total_Estimate_Revenue}
\title{Calculate the Estimated total Revenue}
\usage{
Total_Estimate_Revenue(
  per_of_grocery_spend_rare = 5,
  per_of_grocery_spend_sec = 25,
  per_of_grocery_spend_prim = 60,
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  pct_metro_sec = 40,
  pct_rural_sec = 50,
  pct_town_sec = 50,
  pct_metro_rare = 10,
  pct_rural_rare = 20,
  pct_town_rare = 20,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list,
  state_index,
  est_per_price_increase,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_rare}{the percentage of groceries rare
shoppers will spend at the proposed store}

\item{per_of_grocery_spend_sec}{the percentage of groceries secondary
shoppers will spend at the proposed store}

\item{per_of_grocery_spend_prim}{the percentage of groceries primary
shopper will spend at the proposed store}

\item{pct_metro_prim}{The percentage of primary shoppers in metro market
default taken as 50\% unless specified otherwise.}

\item{pct_rural_prim}{The percentage of primary shoppers in rural market
default taken as 30\% unless specified otherwise.}

\item{pct_town_prim}{The percentage of primary shoppers in town market
default taken as 30\% unless specified otherwise.}

\item{pct_metro_sec}{The percentage of secondary shoppers in metro market
default taken as 40\% unless specified otherwise.}

\item{pct_rural_sec}{The percentage of secondary shoppers in rural market
default taken as 50\% unless specified otherwise.}

\item{pct_town_sec}{The percentage of secondary shoppers in town market
default taken as 50\% unless specified otherwise.}

\item{pct_metro_rare}{The percentage of rare shoppers in metro market
default taken as 10\% unless specified otherwise.}

\item{pct_rural_rare}{The percentage of rare shoppers in rural market
default taken as 20\% unless specified otherwise.}

\item{pct_town_rare}{The percentage of rare shoppers in town market
default taken as 20\% unless specified otherwise.}

\item{county_pop}{population of the county.}

\item{towns_pop}{population of all towns in the county.}

\item{pct_county}{percentage of county in our market.}

\item{metro_list}{lists of population of store location.}

\item{town_list}{list of population of towns in the neighborhood.}

\item{state_index}{the ratio of the current price of the basket to the price of the basket during the base year}

\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
outputs the estimated total revenue at the proposed store
}
\description{
Calculate the Estimated total Revenue
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls Total_Spend_Primary_Shoppers(),
Total_Spend_Secondary_Shoppers() and Total_Spend_Rare_Shoppers()
}
