% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Primary_Shoppers.R
\name{Total_Spend_Primary_Shoppers}
\alias{Total_Spend_Primary_Shoppers}
\title{Calculates the Money Spent by Primary Shoppers}
\usage{
Total_Spend_Primary_Shoppers(
  per_of_grocery_spend_prim = 60,
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  metro_pop,
  town_pop,
  rural_pop,
  state_index,
  est_per_price_increase = 7,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_prim}{The percentage of groceries primary
shopper will spend at the proposed store}

\item{pct_metro_prim}{The percentage of primary shoppers in metro market
default taken as 50\% unless specified otherwise.}

\item{pct_rural_prim}{The percentage of primary shoppers in rural market
default taken as 30\% unless specified otherwise.}

\item{pct_town_prim}{The percentage of primary shoppers in town market
default taken as 30\% unless specified otherwise.}

\item{metro_pop}{Population of the store location}

\item{town_pop}{Population of all towns in the county}

\item{rural_pop}{Population of rural cities in our market}

\item{state_index}{The ratio of the current price of the basket to the price of
the basket during the base year}

\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent
default taken as 7 for 2023.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
Outputs the total spending by primary shoppers at the proposed store
}
\description{
This function return the total money spent by primary shoppers visiting a store.
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Primary_Shopper_Count()
}
\examples{
#Total_Spend_Primary_Shoppers(state_index=99)

}
\seealso{
\code{\link[=Total_Spend_Secondary_Shoppers]{Total_Spend_Secondary_Shoppers()}},\code{\link[=Total_Spend_Rare_Shoppers]{Total_Spend_Rare_Shoppers()}}
}
\author{
Srika Raja
}
