% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Primary_Shoppers.R
\name{Total_Spend_Primary_Shoppers}
\alias{Total_Spend_Primary_Shoppers}
\title{Calculates the money spent by primary shoppers}
\usage{
Total_Spend_Primary_Shoppers(
  per_of_grocery_spend_prim = 60,
  pct_metro_prim = 50,
  pct_rural_prim = 30,
  pct_town_prim = 30,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list,
  state_index,
  est_per_price_increase = 7,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_prim}{the percentage of groceries primary
shopper will spend at the proposed store}

\item{pct_metro_prim}{The percentage of primary shoppers in metro market
default taken as 50\% unless specified otherwise.}

\item{pct_rural_prim}{The percentage of primary shoppers in rural market
default taken as 30\% unless specified otherwise.}

\item{pct_town_prim}{The percentage of primary shoppers in town market
default taken as 30\% unless specified otherwise.}

\item{county_pop}{population of the county.}

\item{towns_pop}{population of all towns in the county.}

\item{pct_county}{percentage of county in our market.}

\item{metro_list}{lists of population of store location.}

\item{town_list}{list of population of towns in the neighborhood.}

\item{state_index}{the ratio of the current price of the basket to the price of the basket during the base year}

\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent
default taken as 7 for 2023.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
outputs the total spending by Primary Shoppers at the proposed store
}
\description{
Calculates the money spent by primary shoppers
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Primary_Shopper_Count()
}
