% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Rare_Shoppers.R
\name{Total_Spend_Rare_Shoppers}
\alias{Total_Spend_Rare_Shoppers}
\title{Calculates the money spent by rare shoppers}
\usage{
Total_Spend_Rare_Shoppers(
  per_of_grocery_spend_rare = 5,
  pct_metro_rare = 10,
  pct_rural_rare = 20,
  pct_town_rare = 20,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list,
  state_index,
  est_per_price_increase,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_rare}{the percentage of groceries rare
shoppers will spend at the proposed store}

\item{pct_metro_rare, pct_rural_rare, pct_town_rare}{The percentage of
rare shoppers in metro,rural and town markets default taken as
10\%,20\%,20\% respectively unless specified otherwise.}

\item{county_pop, towns_pop}{population of the county and population of
all towns in the county}

\item{pct_county}{percentage of county in our market}

\item{metro_list, town_list}{lists of population of store location and
populations of towns in the neighborhood}
}
\value{
outputs the total spending by rare Shoppers at the proposed store
}
\description{
Calculates the money spent by rare shoppers
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Rare_Shopper_Count()
}
