% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Secondary_Shoppers.R
\name{Total_Spend_Secondary_Shoppers}
\alias{Total_Spend_Secondary_Shoppers}
\title{Calculates the money spent by secondary shoppers}
\usage{
Total_Spend_Secondary_Shoppers(
  per_of_grocery_spend_sec = 25,
  pct_metro_sec = 40,
  pct_rural_sec = 50,
  pct_town_sec = 50,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list,
  state_index,
  est_per_price_increase,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_sec}{the percentage of groceries secondary
shoppers will spend at the proposed store}

\item{pct_metro_sec, pct_rural_sec, pct_town_sec}{The percentage of
secondary shoppers in metro,rural and town markets default taken as
40\%,50\%,50\% respectively unless specified otherwise.}

\item{county_pop, towns_pop}{population of the county and population of
all towns in the county}

\item{pct_county}{percentage of county in our market}

\item{metro_list, town_list}{lists of population of store location and
populations of towns in the neighborhood}
}
\value{
outputs the total spending by secondary Shoppers at the proposed store
}
\description{
Calculates the money spent by secondary shoppers
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Secondary_Shopper_Count()
}
