% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Spend_Secondary_Shoppers.R
\name{Total_Spend_Secondary_Shoppers}
\alias{Total_Spend_Secondary_Shoppers}
\title{Calculates the Money Spent by Secondary Shoppers}
\usage{
Total_Spend_Secondary_Shoppers(
  per_of_grocery_spend_sec = 25,
  pct_metro_sec = 40,
  pct_rural_sec = 50,
  pct_town_sec = 50,
  county_pop,
  towns_pop,
  pct_county,
  metro_list,
  town_list,
  state_index,
  est_per_price_increase = 7,
  grocery_sales = 8.11541e+11,
  population = 334233854
)
}
\arguments{
\item{per_of_grocery_spend_sec}{the percentage of groceries secondary
shoppers will spend at the proposed store}

\item{pct_metro_sec}{The percentage of secondary shoppers in metro market
default taken as 40\% unless specified otherwise.}

\item{pct_rural_sec}{The percentage of secondary shoppers in rural market
default taken as 50\% unless specified otherwise.}

\item{pct_town_sec}{The percentage of secondary shoppers in town market
default taken as 50\% unless specified otherwise.}

\item{county_pop}{population of the county}

\item{towns_pop}{population of all towns in the county}

\item{pct_county}{percentage of county in our market}

\item{metro_list}{lists of population of store location}

\item{town_list}{list of population of towns in the neighborhood}

\item{state_index}{the ratio of the current price of the basket to the price of
the basket during the base year}

\item{est_per_price_increase}{Estimated increase in price of grocery in cumulative percent
default taken as 7 for 2023.}

\item{grocery_sales}{Total US grocery sales annually. Default value is 811541000000.}

\item{population}{Total US population. Default value is 334233854.}
}
\value{
outputs the total spending by secondary Shoppers at the proposed store
}
\description{
Calculates the Money Spent by Secondary Shoppers
}
\details{
This function is taken from the estimating Revenue.xlsx and calculated
in step 4. This function calls State_Adj_Capita_Grocery_Spend()
and Secondary_Shopper_Count()
}
\examples{
Total_Spend_Secondary_Shoppers(county_pop=18000,
                               towns_pop=9786,
                               pct_county=17.7,
                               town_list=list(77),
                               metro_list=list(2650),
                               state_index=99,
                               est_per_price_increase=7)

}
