% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTM_Zoner.R
\name{UTM_Zoner}
\alias{UTM_Zoner}
\title{Creates UTM Zones Based on Longitude Values}
\usage{
UTM_Zoner(lng)
}
\arguments{
\item{lng}{longitude}
}
\value{
zone the UTM zone
}
\description{
Allows the user to enter a longitude and have it return the
UTM zone. Do not use this ever, because implementing this is terrible.
}
\author{
Alex Cory
}
